refactor(example): move PlaygroundSettings out of the settings panel
The panel's file opened with 300 lines that were not a panel: two enums and the
immutable settings value class the panel edits. They now live under the
playground's models, beside Employee.

Pure move. The characterisation test passes unchanged, which is the whole
reason it was written first.
